+++ src/cfw/movies/mapper/Mapper.java
@@ -11,7 +11,7 @@ package cfw.movies.mapper;
  * @DATE 2015年8月6日 下午3:10:18
  */
 public interface Mapper<Template> {
-	public Long insertOne(Template t);
+	public int insertOne(Template t);
 	public int deleteOne(Template t);
 	public int updateOne(Template t);
 	
@@ -20,5 +20,5 @@ public interface Mapper<Template> {
 	 * @author Fangwei_Cai
 	 * @time since 2016年3月26日 下午4:59:00
 	 */
-	public Long findOne(Template t);
+	public int findOne(Template t);
 }
